Move MonitorAvailabilityCriteria out of the models package
